requirements.txt
Pillow==8.0.1
pyzbar==0.1.8
